move DataCloudQueryStatus to namespace for client
