+++ jdbc-core/src/test/java/com/salesforce/datacloud/query/v3/DataCloudQueryStatusTest.java
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.salesforce.datacloud.jdbc.core;
+package com.salesforce.datacloud.query.v3;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import com.salesforce.datacloud.jdbc.core.client.DataCloudQueryStatus;
 import java.util.UUID;
 import java.util.function.Consumer;
 import lombok.val;
